♻️Refactor: weather mock data 따로 분리
